+++ cmd/boostd/main.go
@@ -49,6 +49,7 @@ func main() {
 			dagstoreCmd,
 			netCmd,
 			pieceDirCmd,
+			recoverCmd,
 		},
 	}
 	app.Setup()
