+++ cmd/boostd/recover.go
@@ -14,32 +14,24 @@ import (
 
 	"github.com/davecgh/go-spew/spew"
 	"github.com/filecoin-project/boost-gfm/piecestore"
-	piecestoreimpl "github.com/filecoin-project/boost-gfm/piecestore/impl"
-	"github.com/filecoin-project/boost-gfm/storagemarket"
 	"github.com/filecoin-project/boost/cmd/lib"
 	"github.com/filecoin-project/boost/db"
 	"github.com/filecoin-project/boost/piecedirectory"
 	"github.com/filecoin-project/boostd-data/model"
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-commp-utils/writer"
-	vfsm "github.com/filecoin-project/go-ds-versioning/pkg/fsm"
 	"github.com/filecoin-project/go-jsonrpc"
 	"github.com/filecoin-project/go-state-types/abi"
 	"github.com/filecoin-project/go-state-types/builtin/v9/miner"
-	"github.com/filecoin-project/go-statemachine/fsm"
 	"github.com/filecoin-project/lotus/api"
 	"github.com/filecoin-project/lotus/api/v1api"
 	"github.com/filecoin-project/lotus/chain/types"
 	lcli "github.com/filecoin-project/lotus/cli"
-	"github.com/filecoin-project/lotus/lib/backupds"
 	"github.com/filecoin-project/lotus/markets/dagstore"
-	"github.com/filecoin-project/lotus/node/repo"
 	"github.com/filecoin-project/lotus/storage/sealer/storiface"
 	"github.com/google/uuid"
 	"github.com/ipfs/go-cid"
 	"github.com/ipfs/go-cidutil/cidenc"
-	"github.com/ipfs/go-datastore"
-	"github.com/ipfs/go-datastore/namespace"
 	carv2 "github.com/ipld/go-car/v2"
 	"github.com/mitchellh/go-homedir"
 	"github.com/multiformats/go-multibase"
@@ -63,17 +55,17 @@ var (
 	logger *zap.SugaredLogger
 )
 
-var disasterRecoveryCmd = &cli.Command{
-	Name:  "disaster-recovery",
-	Usage: "Disaster Recovery commands",
+var recoverCmd = &cli.Command{
+	Name:  "recover",
+	Usage: "LID recover commands",
 	Subcommands: []*cli.Command{
-		restorePieceStoreCmd,
+		lidCmd,
 	},
 }
 
-var restorePieceStoreCmd = &cli.Command{
-	Name:   "restore-piece-store",
-	Usage:  "Restore Piece store",
+var lidCmd = &cli.Command{
+	Name:   "lid",
+	Usage:  "lid",
 	Before: before,
 	Flags: []cli.Flag{
 		&cli.StringFlag{
@@ -92,9 +84,9 @@ var restorePieceStoreCmd = &cli.Command{
 			//Required: true,
 		},
 		&cli.StringFlag{
-			Name:  "disaster-recovery-dir",
-			Usage: "location to store progress of disaster recovery",
-			Value: "~/.boost-disaster-recovery",
+			Name:  "recover-dir",
+			Usage: "location to store progress of recovery",
+			Value: "~/.boost-recover",
 		},
 		&cli.StringFlag{
 			Name:  "repo",
@@ -122,7 +114,7 @@ var restorePieceStoreCmd = &cli.Command{
 		},
 	},
 	Action: func(cctx *cli.Context) error {
-		c := make(chan os.Signal)
+		c := make(chan os.Signal, 1)
 		errc := make(chan error)
 
 		signal.Notify(c, os.Interrupt, syscall.SIGTERM)
@@ -154,7 +146,7 @@ func action(cctx *cli.Context) error {
 	ctx := lcli.ReqContext(cctx)
 
 	var err error
-	dr, err = NewDisasterRecovery(ctx, cctx.String("disaster-recovery-dir"), cctx.String("repo"))
+	dr, err = NewDisasterRecovery(ctx, cctx.String("recover-dir"), cctx.String("repo"))
 	if err != nil {
 		return err
 	}
@@ -162,7 +154,7 @@ func action(cctx *cli.Context) error {
 	var sectorid abi.SectorNumber
 	if cctx.IsSet("sector-id") {
 		sectorid = abi.SectorNumber(cctx.Uint64("sector-id"))
-		logger.Infow("running disaster recovery tool on a single sector", "sector", sectorid)
+		logger.Infow("running recovery tool on a single sector", "sector", sectorid)
 	}
 
 	ignoreCommp = cctx.Bool("ignore-commp")
@@ -270,7 +262,7 @@ func action(cctx *cli.Context) error {
 }
 
 type DisasterRecovery struct {
-	Dir     string // main disaster recovery dir - keeps progress on recovery
+	Dir     string // main recovery dir - keeps progress on recovery
 	DoneDir string
 
 	TotalSectors int
@@ -309,10 +301,10 @@ type PieceStatus struct {
 func NewDisasterRecovery(ctx context.Context, dir, repodir string) (*DisasterRecovery, error) {
 	drDir, err := homedir.Expand(dir)
 	if err != nil {
-		return nil, fmt.Errorf("expanding disaster recovery dir path: %w", err)
+		return nil, fmt.Errorf("expanding recovery dir path: %w", err)
 	}
 	if drDir == "" {
-		return nil, errors.New("disaster-recovery-dir is a required flag")
+		return nil, errors.New("recover-dir is a required flag")
 	}
 
 	repodir, err = homedir.Expand(repodir)
@@ -356,7 +348,7 @@ func NewDisasterRecovery(ctx context.Context, dir, repodir string) (*DisasterRec
 	}
 
 	if recoverRanPreviously {
-		logger.Warn("disaster recovery directory exists, so will continue from where recovery left off perviously")
+		logger.Warn("recovery directory exists, so will continue from where recovery left off previously")
 	}
 
 	err = drr.loadPieceStoreAndBoostDB(ctx, repodir)
@@ -372,7 +364,7 @@ func NewDisasterRecovery(ctx context.Context, dir, repodir string) (*DisasterRec
 
 func (dr *DisasterRecovery) loadPieceStoreAndBoostDB(ctx context.Context, repoDir string) error {
 	// Open the datastore in the existing repo
-	ds, err := openDataStore(repoDir)
+	ds, err := lib.OpenDataStore(repoDir)
 	if err != nil {
 		return fmt.Errorf("creating piece store from repo %s: %w", repoDir, err)
 	}
@@ -380,12 +372,12 @@ func (dr *DisasterRecovery) loadPieceStoreAndBoostDB(ctx context.Context, repoDi
 	// Create a mapping of on-chain deal ID to deal proposal cid.
 	// This is needed below so that we can map from the legacy piece store
 	// info to a legacy deal.
-	dr.PropCidByChainDealID, err = getPropCidByChainDealID(ctx, ds)
+	dr.PropCidByChainDealID, err = lib.GetPropCidByChainDealID(ctx, ds)
 	if err != nil {
 		return fmt.Errorf("building chain deal id -> proposal cid map: %w", err)
 	}
 
-	ps, err = openPieceStore(ctx, ds)
+	ps, err = lib.OpenPieceStore(ctx, ds)
 	if err != nil {
 		return fmt.Errorf("opening piece store: %w", err)
 	}
@@ -605,7 +597,7 @@ func processPiece(ctx context.Context, sectorid abi.SectorNumber, chainDealID ab
 		var shouldGenerateNewDeal bool
 
 		if dr.HaveBoostDealsAndPieceStore { // successfully loaded boost sqlite db and piece store => try to infer dealinfo
-			// Find the deal corresponding to the deal info's DealID
+			// find the deal corresponding to the deal info's DealID
 			proposalCid, okLegacy := dr.PropCidByChainDealID[chainDealID]
 			uuid, okBoost := dr.BoostDeals[chainDealID]
 
@@ -641,7 +633,7 @@ func processPiece(ctx context.Context, sectorid abi.SectorNumber, chainDealID ab
 		}
 
 		if !dr.HaveBoostDealsAndPieceStore || shouldGenerateNewDeal { // missing boost sqlite db and piece store, so generate new dealinfo
-			//TODO: regenerate boost db sqlite??
+			// in the future we could also regenerate boost db sqlite??
 
 			di := model.DealInfo{
 				DealUuid:    uuid.NewString(),
@@ -649,7 +641,7 @@ func processPiece(ctx context.Context, sectorid abi.SectorNumber, chainDealID ab
 				ChainDealID: chainDealID,
 				MinerAddr:   maddr,
 				SectorID:    sectorid,
-				PieceOffset: offset.Padded(), // TODO: confirm that this is correct...?
+				PieceOffset: offset.Padded(),
 				PieceLength: piecesize,
 			}
 
@@ -786,111 +778,6 @@ type Reader interface {
 	io.Reader
 	io.ReaderAt
 	io.Seeker
-}
-
-func openDataStore(path string) (*backupds.Datastore, error) {
-	ctx := context.Background()
-
-	rpo, err := repo.NewFS(path)
-	if err != nil {
-		return nil, fmt.Errorf("could not open repo %s: %w", path, err)
-	}
-
-	exists, err := rpo.Exists()
-	if err != nil {
-		return nil, fmt.Errorf("checking repo %s exists: %w", path, err)
-	}
-	if !exists {
-		return nil, fmt.Errorf("repo does not exist: %s", path)
-	}
-
-	lr, err := rpo.Lock(repo.StorageMiner)
-	if err != nil {
-		return nil, fmt.Errorf("locking repo %s: %w", path, err)
-	}
-
-	mds, err := lr.Datastore(ctx, "/metadata")
-	if err != nil {
-		return nil, err
-	}
-
-	bds, err := backupds.Wrap(mds, "")
-	if err != nil {
-		return nil, fmt.Errorf("opening backupds: %w", err)
-	}
-
-	return bds, nil
-}
-
-func getPropCidByChainDealID(ctx context.Context, ds *backupds.Datastore) (map[abi.DealID]cid.Cid, error) {
-	deals, err := getLegacyDealsFSM(ctx, ds)
-	if err != nil {
-		return nil, err
-	}
-
-	// Build a mapping of chain deal ID to proposal CID
-	var list []storagemarket.MinerDeal
-	if err := deals.List(&list); err != nil {
-		return nil, err
-	}
-
-	byChainDealID := make(map[abi.DealID]cid.Cid, len(list))
-	for _, d := range list {
-		if d.DealID != 0 {
-			byChainDealID[d.DealID] = d.ProposalCid
-		}
-	}
-
-	return byChainDealID, nil
-}
-
-func openPieceStore(ctx context.Context, ds *backupds.Datastore) (piecestore.PieceStore, error) {
-	// Open the piece store
-	ps, err := piecestoreimpl.NewPieceStore(namespace.Wrap(ds, datastore.NewKey("/storagemarket")))
-	if err != nil {
-		return nil, fmt.Errorf("creating piece store from datastore : %w", err)
-	}
-
-	// Wait for the piece store to be ready
-	ch := make(chan error, 1)
-	ps.OnReady(func(e error) {
-		ch <- e
-	})
-
-	err = ps.Start(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("starting piece store: %w", err)
-	}
-
-	select {
-	case err = <-ch:
-		if err != nil {
-			return nil, fmt.Errorf("waiting for piece store to be ready: %w", err)
-		}
-	case <-ctx.Done():
-		return nil, errors.New("cancelled while waiting for piece store to be ready")
-	}
-
-	return ps, nil
-}
-
-func getLegacyDealsFSM(ctx context.Context, ds *backupds.Datastore) (fsm.Group, error) {
-	// Get the deals FSM
-	provDS := namespace.Wrap(ds, datastore.NewKey("/deals/provider"))
-	deals, migrate, err := vfsm.NewVersionedFSM(provDS, fsm.Parameters{
-		StateType:     storagemarket.MinerDeal{},
-		StateKeyField: "State",
-	}, nil, "2")
-	if err != nil {
-		return nil, fmt.Errorf("reading legacy deals from datastore: %w", err)
-	}
-
-	err = migrate(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("running provider fsm migration script: %w", err)
-	}
-
-	return deals, err
 }
 
 func createLogger(logPath string) (*zap.SugaredLogger, error) {
