+++ cmd/boostx/main.go
@@ -38,7 +38,6 @@ func main() {
 			marketWithdrawCmd,
 			statsCmd,
 			sectorCmd,
-			disasterRecoveryCmd,
 		},
 	}
 	app.Setup()
@@ -50,7 +49,6 @@ func main() {
 
 func before(cctx *cli.Context) error {
 	_ = logging.SetLogLevel("boostx", "INFO")
-	//_ = logging.SetLogLevel("sectoraccessor", "DEBUG")
 
 	if cliutil.IsVeryVerbose {
 		_ = logging.SetLogLevel("boostx", "DEBUG")
